+++ src/dipdup/context.py
@@ -408,17 +408,18 @@ class CallbackManager:
         *args,
         **kwargs: Any,
     ) -> None:
+        module = f'{self._package}.handlers.{name}'
         handler_config = self._get_handler(name, index)
         new_ctx = HandlerContext(
             datasources=ctx.datasources,
             config=ctx.config,
             callbacks=ctx.callbacks,
-            logger=FormattedLogger(f'dipdup.handlers.{name}', fmt),
+            logger=FormattedLogger(module, fmt),
             handler_config=handler_config,
             datasource=datasource,
         )
         # NOTE: Handlers are not atomic, levels are. Do not open transaction here.
-        with self._callback_wrapper('handler', name):
+        with self._callback_wrapper(module):
             await handler_config.callback_fn(new_ctx, *args, **kwargs)
 
     async def fire_hook(
@@ -430,12 +431,13 @@ class CallbackManager:
         *args,
         **kwargs: Any,
     ) -> None:
+        module = f'{self._package}.hooks.{name}'
         hook_config = self._get_hook(name)
         new_ctx = HookContext(
             datasources=ctx.datasources,
             config=ctx.config,
             callbacks=ctx.callbacks,
-            logger=FormattedLogger(f'dipdup.hooks.{name}', fmt),
+            logger=FormattedLogger(module, fmt),
             hook_config=hook_config,
         )
 
@@ -443,8 +445,7 @@ class CallbackManager:
 
         async def _wrapper():
             async with AsyncExitStack() as stack:
-
-                stack.enter_context(self._callback_wrapper('hook', name))
+                stack.enter_context(self._callback_wrapper(module))
                 if hook_config.atomic:
                     await stack.enter_async_context(in_global_transaction())
 
@@ -471,16 +472,16 @@ class CallbackManager:
         await execute_sql_scripts(connection, sql_path)
 
     @contextmanager
-    def _callback_wrapper(self, kind: str, name: str) -> Iterator[None]:
+    def _callback_wrapper(self, module: str) -> Iterator[None]:
         try:
             with ExitStack() as stack:
                 if Metrics.enabled:
-                    stack.enter_context(Metrics.measure_callback_duration(name))
+                    stack.enter_context(Metrics.measure_callback_duration(module))
                 yield
         except Exception as e:
             if isinstance(e, ReindexingRequiredError):
                 raise
-            raise CallbackError(kind, name) from e
+            raise CallbackError(module, e) from e
 
     @classmethod
     def _verify_arguments(cls, ctx: HookContext, *args, **kwargs) -> None: