+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 * context: Fixed reporting incorrect reindexing reason.
+* exceptions: Fixed crash with `FrozenInstanceError` when exception is raised from callback.
 
 ### Improved